pkg/ioutilx: move UTF-16 helpers to pkg/strutil
Keep ReadAtMaximum in ioutilx (io-like). Move FromUTF16le /
FromUTF16leToString to pkg/strutil so ioutilx only holds ioutil-style
helpers. Updates the executil Windows UTF-16 command helper.

Use gotest.tools/v3/assert in the new strutil UTF-16 tests instead of
t.Fatal to satisfy forbidigo.

Fixes #5283
