+++ pkg/ioutilx/ioutilx.go
@@ -7,9 +7,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-
-	"golang.org/x/text/encoding/unicode"
-	"golang.org/x/text/transform"
 )
 
 // ReadAtMaximum reads n at maximum.
@@ -25,22 +22,4 @@ func ReadAtMaximum(r io.Reader, n int64) ([]byte, error) {
 		}
 	}
 	return b, err
-}
-
-// FromUTF16le returns an io.Reader for UTF16le data.
-// Windows uses little endian by default, use unicode.UseBOM policy to retrieve BOM from the text,
-// and unicode.LittleEndian as a fallback.
-func FromUTF16le(r io.Reader) io.Reader {
-	o := transform.NewReader(r, unicode.UTF16(unicode.LittleEndian, unicode.UseBOM).NewDecoder())
-	return o
-}
-
-// FromUTF16leToString reads from Unicode 16 LE encoded data from an io.Reader and returns a string.
-func FromUTF16leToString(r io.Reader) (string, error) {
-	out, err := io.ReadAll(FromUTF16le(r))
-	if err != nil {
-		return "", err
-	}
-
-	return string(out), nil
 }
