+++ pkg/strutil/strutil.go
@@ -1,0 +1,28 @@
+// SPDX-FileCopyrightText: Copyright The Lima Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package strutil
+
+import (
+	"io"
+
+	"golang.org/x/text/encoding/unicode"
+	"golang.org/x/text/transform"
+)
+
+// FromUTF16le returns an io.Reader for UTF16le data.
+// Windows uses little endian by default, use unicode.UseBOM policy to retrieve BOM from the text,
+// and unicode.LittleEndian as a fallback.
+func FromUTF16le(r io.Reader) io.Reader {
+	return transform.NewReader(r, unicode.UTF16(unicode.LittleEndian, unicode.UseBOM).NewDecoder())
+}
+
+// FromUTF16leToString reads from Unicode 16 LE encoded data from an io.Reader and returns a string.
+func FromUTF16leToString(r io.Reader) (string, error) {
+	out, err := io.ReadAll(FromUTF16le(r))
+	if err != nil {
+		return "", err
+	}
+
+	return string(out), nil
+}
